fix(types): resolve cascading mypy errors from Pos type widening
- engine.py: use Pos type alias for state_before_summary, remove unused imports
- learn.py: widen by_action type to tuple[int|float, int|float], fix motion_by_action mode computation
- fallback.py: use float casts for pos coordinates, int() return for Manhattan distance
- heuristics.py: add int() cast for reach_radius magnitude computation
- counter_evidence.py: restore Any import, use Pos type alias
- __init__.py: ruff import sort

+++ planning/fallback.py
@@ -53,9 +53,9 @@ def pick_fallback_unknown(
             return 0
         for _eid, (dim, val) in ua.state.relevant:
             if dim == "pos" and isinstance(val, tuple):
-                dr: int = int(val[0]) - current[0]
-                dc: int = int(val[1]) - current[1]
-                return abs(dr) + abs(dc)
+                dr: float = float(val[0]) - current[0]
+                dc: float = float(val[1]) - current[1]
+                return int(abs(dr) + abs(dc))
         return 0
 
     return min(fresh, key=_dist)
